rush-apply-agent/requirements.txt
playwright>=1.44.0
requests>=2.31.0
python-dotenv>=1.0.0
reportlab>=4.2.0
